build 12.2.13-noggalito version
we override `#addFieldToFormData` method so that it doesn’t add null values to the form data object to upload to the server.

when we serialise `null` values and post them to the server using `Content-Type: multipart/form`, they’re sent in plain text - so null is the same as “null”. And our server expects JSON so it mistakenly overrides our data.